+++ eyes_core/applitools/core/feature.py
@@ -1,0 +1,5 @@
+from enum import Enum
+
+
+class Feature(Enum):
+    TARGET_WINDOW_CAPTURES_SELECTED_FRAME = "Target window captures selected frame"